tensorflow: fix pyconfig-target.h include nested
While do_configure failed for other reason and rerun again,
pyconfig-target.h will be duplicated with pyconfig.h:
...
pyconfig-target.h:2:29: error: #include nested too deeply
...

Test the existence of pyconfig-target.h before moving
could workaround the issue.

+++ recipes-framework/tensorflow/tensorflow_1.13.0.bb
@@ -70,8 +70,9 @@ TF_CONFIG ?= " \
 "
 do_configure_append () {
     CROSSTOOL_PYTHON_INCLUDE_PATH="${STAGING_INCDIR}/python${PYTHON_BASEVERSION}${PYTHON_ABI}"
-    install -d ${CROSSTOOL_PYTHON_INCLUDE_PATH}
-    mv ${CROSSTOOL_PYTHON_INCLUDE_PATH}/pyconfig.h ${CROSSTOOL_PYTHON_INCLUDE_PATH}/pyconfig-target.h
+    if [ ! -e ${CROSSTOOL_PYTHON_INCLUDE_PATH}/pyconfig-target.h ];then
+        mv ${CROSSTOOL_PYTHON_INCLUDE_PATH}/pyconfig.h ${CROSSTOOL_PYTHON_INCLUDE_PATH}/pyconfig-target.h
+    fi
 
     install -m 644 ${STAGING_INCDIR_NATIVE}/python${PYTHON_BASEVERSION}${PYTHON_ABI}/pyconfig.h \
         ${CROSSTOOL_PYTHON_INCLUDE_PATH}/pyconfig-native.h
